Lynx

Documentation for Lynx.

Lynx.AbstractLayout

Lynx.App

Lynx.SideBar

Lynx.SideBar

Lynx.Box

Lynx.Button

Lynx.CanvasOnly

Lynx.CheckBox

Lynx.ColorButton

Lynx.ComboBox

Lynx.ComboBoxText

Lynx.FileChooserDialog

Lynx.Frame

Lynx.GLArea

Lynx.Image

Lynx.Label

Lynx.LinkButton

Lynx.MessageDialog

Lynx.Notebook

Lynx.Paned

Lynx.ProgressBar

Lynx.Scrolled

Lynx.Slider

Lynx.SpinButton

Lynx.Spinner

Lynx.Switch

Lynx.TextField

Lynx.ToggleBtn

Lynx.VolumeBtn

Lynx.Window

Lynx.framerate!

Lynx.gcpreserve

Lynx.getprop

Lynx.layout!

Lynx.loop!

Lynx.mapr

Lynx.mapr

Lynx.onmousedown

Lynx.onmousedrag

Lynx.onmousemove

Lynx.onmouseup

Lynx.onscroll

Lynx.run!

Lynx.setprop!

Lynx.use!

Lynx.@app

Lynx.@canvas

Lynx.@framerate

Lynx.@on

Lynx.@use

Lynx.@waitfor

Lynx.@window

Lynx.AbstractLayout — Type

Layouts are used to organize the canvas and other widgets within the window.

They are simple function or callable structures.

Example of layout implementation:

struct MyLayout <: AbstractLayout
    fields...
    function MyLayout(args...)
        ...
    end
end

function (layout::MyLayout)(window, canvas, widgets...)
    ...
end

The call implementation of a layout should return a widget that contains (or not) the canvas and the widgets. The returned widget will then be added to the window.

Lynx.App — Type

An App is a basic structure that holds a window, a canvas and other widgets to make a interactive application with a few lines of code.

Lynx.SideBar — Type

The SideBar layout divides the window between the canvas and a sidebar containing the other widgets.

Lynx.SideBar — Type

    SideBar(position = :left, size = .25, props = NamedTuple()) -> SideBar

Creates a SideBar layout. The position determines where the sidebar is placed relative to the canvas (can be either :left or :right). The size controls the width of the sidebar relative to the width of the window (must be a value between 0 and 1). You can also pass aditional properties for the sidebar widget using the props parameter.

Lynx.Box — Method

    Box(layout; props...) -> Widget{GtkBox}

A container for packing widgets in a single row or column

Lynx.Button — Method

    Button(text; props...) -> Widget{GtkButton}

A widget that emits a signal when clicked on.

Lynx.CanvasOnly — Method

The canvas is the only widget and fills the entire window.

Lynx.CheckBox — Method

    CheckBox(; props...) -> Widget{GtkCheckButton}

A widget with with a check box.

Lynx.ColorButton — Method

    ColorButton(color; props...) -> Widget{GtkColorButton}

A button to launch a color selection dialog

Lynx.ComboBox — Method

    ComboBox(args...; props...) -> Widget{GtkComboBox}

A widget used to choose from a list of items

Lynx.ComboBoxText — Method

ComboBoxText(args...; props...) -> Widget{GtkComboBoxText}

A simple, text-only combo box

Lynx.FileChooserDialog — Method

    FileChooserDialog(args...; props...) -> Widget{GtkFileChooserDialog}

A file chooser dialog, suitable for “File/Open” or “File/Save” commands

Lynx.Frame — Method

    Frame(args...; props...) -> Widget{GtkFrame}

A bin with a decorative frame and optional label

Lynx.GLArea — Method

    GLArea(args...; props...) -> Widget{GtkGLArea}

A widget for custom drawing with OpenGL

Lynx.Image — Method

    Image([filepath]; props...) --> Widget{GtkImage}

A widget displaying an image

Lynx.Label — Method

    Label(text; props...) -> Widget{GtkLabel}

A widget that displays a small to medium amount of text

Lynx.LinkButton — Method

    LinkButton(text; props...) -> Widget{GtkLinkButton}

Create buttons bound to a URL

Lynx.MessageDialog — Method

    MessageDialog(args...; props...) -> Widget{GtkMessageDialog}

A convenient message window

Lynx.Notebook — Method

    Notebook(args...; props...) -> Widget{GtkNotebook}

A tabbed notebook container

Lynx.Paned — Method

    Paned(layout; props...) -> Widget{GtkPaned}

A widget with two adjustable panes

Lynx.ProgressBar — Method

    ProgressBar(; props...) -> Widget{GtkProgressBar}

A widget which indicates progress visually

Lynx.Scrolled — Method

    Scrolled(args...; props...) -> Widget{GtkScrolledWindow}

Adds scrollbars to its child widget

Lynx.Slider — Function

    Slider(range, vertical = false; props...) -> Widget{GtkScale}

A slider widget for selecting a value from a range

Lynx.SpinButton — Method

    SpinButton(range; props...) -> Widget{GtkSpinButton}

Retrieve an integer or floating-point number from the user

Lynx.Spinner — Method

    Spinner(; props...) -> Widget{GtkSpinner}

Show a spinner animation

Lynx.Switch — Method

    Switch(; props...) -> Widget{GtkSwitch}

A “light switch” style toggle

Lynx.TextField — Method

    TextField(; props...) -> Widget{GtkEntry}

A single line text entry field

Lynx.ToggleBtn — Method

    ToggleBtn(; props...) -> Widget{GtkToggleButton}

Create buttons which retain their state

Lynx.VolumeBtn — Method

    VolumeBtn(; props...) -> Widget{GtkVolumeButton}

A button which pops up a volume control

Lynx.Window — Method

    Window(title, width, height; props...) -> Widget{GtkWindow}

Toplevel which can contain other widgets

Lynx.framerate! — Method

framerate!(x::Real)

Sets the current framerate.

Lynx.gcpreserve — Method

gcpreserve(widget::GtkWidget, obj)

Preserve obj until widget has been destroyed.

Lynx.getprop — Method

    getprop(widget, name, T) -> T

Gets a widget property indentified by name with type T.

Lynx.layout! — Method

layout!(layout::Function)

Sets the current layout.

Lynx.loop! — Method

loop!(x::Bool)

Whether or not the update function should be called every frame.

Lynx.mapr — Method

    mapr(t, a, b, c, d) -> Real

Maps the value t from the interval [a, b] to [c, d]

Lynx.mapr — Method

    mapr(t, a, b) -> Real

Maps the value t from the interval [0, 1] to [a, b]

Lynx.onmousedown — Method

    onmousedown(callback, canvas)

The mouse-down event is fired when the user press with the mouse on the canvas.

Lynx.onmousedrag — Method

    onmousedrag(callback, canvas)

The mouse-drag event is fired when the user holds down and move the cursor on the canvas.

Lynx.onmousemove — Method

    onmousemove(callback, canvas)

The mouse-move event is fired when the user moves the cursor on the canvas.

Lynx.onmouseup — Method

    onmouseup(callback, canvas)

The mouse-up event is fired when the user release the mouse on the canvas.

Lynx.onscroll — Method

    onscroll(callback, canvas)

The scroll event is fired when the user scrolls with the mouse-wheel on the canvas.

Lynx.run! — Method

    run!(update::Function; await = false, hotreload = false)

Run the current App with the given update function. If await is true, the main task will be blocked until the user closes the window. If hotreload is true, only the newest version of update will be called.

Lynx.setprop! — Method

    setprop!(widget, name, value)

Sets the value of a widget property indentified by name.

Lynx.use! — Method

    use!(widget) -> AbstractWidget

Adds the widget to the current application and returns it.

Lynx.@app — Macro

@app() -> App

Returns the current App or throws an exception if no App has been created.

Lynx.@canvas — Macro

@canvas() -> Canvas

Returns the current Canvas.

Lynx.@framerate — Macro

@framerate() -> Float64

Returns the current framerate.

Lynx.@on — Macro

    @on widget.event() do args... body... end

Adds a event callback to the corresponding widget

Usage

@on window.destroy() do args...
    # do something when the window is destroyed
end

@on button.clicked() do args...
    # do something when the button is clicked
end

Lynx.@use — Macro

    @use widget

Adds the widget to the current application and returns it.

Lynx.@waitfor — Macro

    @waitfor widget.event

Block the current task until the event is triggered for widget.

Examples:

window = Window("Hello, world", 800, 600)

... # Some UI stuff

if !isinteractive()
    # If the code is not running on a interactive julia session
    # (e.g. the REPL), then the line below will block the current
    # task until the window is destroyed
    @waitfor window.destroy
end

Lynx.@window — Macro

@window() -> Widget{GtkWindow}

Returns the current Window.